Added little debug code to eventhandler. MouseLeft creates enemy and mouseleft release destroys it. You can move enemy on screen if you keep mouseleft pressed and test projectile code.

+++ src/GameLogic/cTowerEntity.h
@@ -67,7 +67,7 @@ private:
     unsigned int range_;
 
     /// Reload timeout
-    int reloadTimeout_;
+    float reloadTimeout_;
 
     /// Last shot timestamp
     float lastShotTime_;
